feat: add share dialog (#5547)
* feat: add share dialog

* fix: add cspell word

* Update packages/mask/src/plugins/Pets/constants.ts


* fix: error word

* fix: show in profile

* fix: change word notice

* fix: useTimeout
