+++ cspell.json
@@ -325,6 +325,7 @@
         "nftred",
         "nftx",
         "noamount",
+        "nonfungiblefriends",
         "northcutt",
         "notistack",
         "nounset",
