+++ packages/mask/shared-ui/locales/en-US.json
@@ -896,11 +896,13 @@
     "plugin_furucombo_head_pools": "Pools",
     "plugin_furucombo_head_action": "Action",
     "plugin_pets_dialog_title": "Non-Fungible Friends",
+    "plugin_pets_dialog_title_share": "Successful",
     "plugin_pets_dialog_contract": "NFT Contract",
     "plugin_pets_dialog_token": "Token ID",
     "plugin_pets_dialog_msg": "Greeting message",
     "plugin_pets_dialog_msg_optional": "Optional, 100 characters max.",
     "plugin_pets_dialog_btn": "Confirm",
+    "plugin_pets_dialog_btn_share": "Share",
     "plugin_pets_dialog_preview": "Preview",
     "plugin_pets_dialog_created": "Created by MintTeam",
     "plugin_pets_dialog_powered": "Powered by RSS3",
