+++ packages/mask/src/plugins/Pets/constants.ts
@@ -6,7 +6,8 @@ export const PetsPluginID = 'com.maskbook.pets'
 export const TWITTER = 'twitter.com'
 export const MASK_TWITTER = 'realMaskNetwork'
 export const NFTS_CONFIG_ADDRESS = '0x500cF2ea6755ea35eAC8727A70D997ecc3d26258'
-export const DEFAULT_SET_WORD = 'Click the Mask icon to the left and set up your Non-Fungible Friend in "NFTs" tab!'
+export const DEFAULT_SET_WORD =
+    'Click the wallet icon on the left side of the page and set up your Non-Fungible Friend in "NFTs" tab!'
 export const DEFAULT_PUNK_MASK_WORD =
     "I'm CryptoPunk #6128... in 3D & with a body! Voyagers, welcome to the uncharted waters of WEB3!"
 
@@ -33,6 +34,11 @@ export const Punk3D = {
         type: 'model/gltf-binary',
     }),
 }
+
+export const Share_Twitter = 'https://twitter.com/realMaskNetwork/status/1486648872558424064'
+export const Share_Twitter_TXT = encodeURIComponent(
+    `I just set up my Non-Fungible Friend with @realMaskNetwork (powered by @MintTeamNFT). Visit my profile to check it out! Install Mask Network extension from mask.io and set yours.\n #mask_io #nonfungiblefriends\n${Share_Twitter}`,
+)
 
 export const GLB3DIcon = new URL('./assets/glb3D.png', import.meta.url).toString()
 export const CloseIcon = new URL('./assets/close.png', import.meta.url).toString()
