+++ packages/mask/src/plugins/Pets/hooks/useEssay.ts
@@ -4,6 +4,7 @@ import { useAsync } from 'react-use'
 import { PluginPetRPC } from '../messages'
 import type { User } from '../types'
 import { Punk3D, DEFAULT_SET_WORD, MASK_TWITTER, DEFAULT_PUNK_MASK_WORD, PunkIcon } from './../constants'
+import { useUser } from './useUser'
 
 export function useEssay(user: User, refresh?: number) {
     return useAsync(async () => {
@@ -15,16 +16,19 @@ export function useEssay(user: User, refresh?: number) {
 
 export function useDefaultEssay(user: User) {
     const [essayMeta, setEssayMeta] = useState<ShowMeta | undefined>(undefined)
+    const profileUser = useUser()
     useEffect(() => {
         if (user?.userId || user?.userId !== '$unknown') {
+            const isProfile = user.userId === profileUser.userId
+            const isMASK = user.userId === MASK_TWITTER
             setEssayMeta({
-                image: user.userId === MASK_TWITTER ? Punk3D.url : PunkIcon,
-                word: user.userId === MASK_TWITTER ? DEFAULT_PUNK_MASK_WORD : DEFAULT_SET_WORD,
-                type: user.userId === MASK_TWITTER ? ImageType.GLB : ImageType.NORMAL,
+                image: isMASK ? Punk3D.url : isProfile ? PunkIcon : '',
+                word: isMASK ? DEFAULT_PUNK_MASK_WORD : DEFAULT_SET_WORD,
+                type: isMASK ? ImageType.GLB : ImageType.NORMAL,
             })
         } else {
             setEssayMeta(undefined)
         }
-    }, [user])
+    }, [user, profileUser])
     return essayMeta
 }
